ET-3501 move ingestion endpoints (part 1) (11) (#685)
Moves the new_documents endpoint to actix.

Split out from the rest as it also merges the two different bulk request implementations.

**References:**

- [ET-3501]
- based on:
  - #684
  - #674
  - #673   
  - #672 
  - #671   



[ET-3501]: https://xainag.atlassian.net/browse/ET-3501?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

+++ discovery_engine_core/web-api2/src/ingestion.rs
@@ -64,7 +64,7 @@ pub struct IngestionConfig {
     #[allow(dead_code)]
     #[as_ref]
     #[serde(default = "default_max_document_batch_size")]
-    pub(crate) max_document_batch_size: u64,
+    pub(crate) max_document_batch_size: usize,
 }
 
 impl Default for IngestionConfig {
@@ -75,7 +75,7 @@ impl Default for IngestionConfig {
     }
 }
 
-fn default_max_document_batch_size() -> u64 {
+fn default_max_document_batch_size() -> usize {
     100
 }
